chore: check if the key exists before removing it

+++ clis/shipper/json.go
@@ -26,6 +26,10 @@ func GetJSON(r HTTPRequest) (*gabs.Container, error) {
 
 // DeleteJSONKey deletes a key from the JSON object
 func DeleteJSONKey(json *gabs.Container, key string) {
+	if !json.Exists(key) {
+		return
+	}
+
 	err := json.DeleteP(key)
 	if err != nil {
 		fmt.Printf(">> cannot delete %s from JSON object: %s\n", key, err)
